+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased changes
 
+- Add an `update` option to `HtmlQuery.table`, which is a function that is called for every table cell and updates
+  each cell's contents.
+
 ## 4.2.1
 
 - Add sponsorship link.